// backend/src/types.ts
export type Identifier = string

export type Entity = eObject | eAnimation | eLayer

export type eObject = {
  type: EntityTypes.object
  parameters: ObjectParameters & Shape
}

// Note: If objects are in multiple layers, most recently applied animation → layer should take precedence
export type eLayer = {
  type: EntityTypes.layer
  parameters: LayerParameters
}

export enum LayerParametersEnum {
  priority = 'priority',
  fill = 'fill',
  x = 'x',
  y = 'y',
  opacity = 'opacity',
  shape = 'shape',
  radius = 'radius',
  width = 'width',
  height = 'height',
  points = 'points',
  length = 'length',
  objects = 'objects',
}

type LayerParameters = {
  priority?: number
  fill?: string
  x?: number
  y?: number
  opacity?: number
  shape?: string
  radius?: number
  width?: number
  height?: number
  points?: Array<[number, number]>
  length?: number
  objects?: Identifier[]
}

export enum ObjectParametersEnum {
  priority = 'priority',
  fill = 'fill',
  x = 'x',
  y = 'y',
  opacity = 'opacity',
  shape = 'shape',
  radius = 'radius',
  width = 'width',
  height = 'height',
  points = 'points',
  length = 'length',
}

type ObjectParameters = {
  priority?: number
  style?: Partial<Style>
}

type Style = {
  fill?: string
  x?: number
  y?: number
  opacity?: number
}

type Shape = Circle | Rectangle | Triangle | Pentagon | Hexagon | Octagon | Polygon
// ...
export type eAnimation = {
  type: EntityTypes.animation
  parameters: {
    duration?: number
    delay?: number
    fill?: string
    opacity?: number
  } & AnimationVariant
}

type AnimationVariant = Translation | Rotation | Scaling

type Translation = {
  type: 'translation'
  velocity: number
  direction: Direction | Vector2D
}

type Rotation = {
  type: 'rotation'
  angularVelocity: number
}

type Scaling = {
  type: 'scale'
  scaleFactor: number
}

type Vector2D = [number, number]

export enum Direction {
  Up,
  Down,
  Left,
  Right,
}

// -- Execution --

export enum EntityTypes {
  object = 'object',
  layer = 'layer',
  animation = 'animation',
}

export type Execution = Step[]

export type Step = Persistence | Delete | Animate | Global | Conditional

// --

export type Persistence = Create | Update

export type Parameter = { parameter: string; value: any }

export type Parameters =
  | Partial<eObject['parameters']>
  | Partial<eAnimation['parameters']>
  | Partial<eLayer['parameters']>

export type Create = {
  type: 'create'
  identifier: string
  entity: EntityTypes
  parameters: Parameters
}

export type Update = {
  type: 'update'
  identifier: string
  entity: EntityTypes
  parameters: Parameters
}